Refactor TOC and quiz scripts into modules
